# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API De Tuna Resort",
        "description": "Dokumentasi API untuk Sistem Informasi & Pemesanan De Tuna Resort",
        "contact": {
            "email": "admin@detuna.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server API De Tuna (Lokal - php artisan serve)"
        },
        {
            "url": "http://detuna-api.test",
            "description": "Server API De Tuna (Lokal - Laragon)"
        }
    ],
    "paths": {
...
        "/api/v1/admin/resort-facilities": {
            "get": {
                "tags": [
                    "Admin - Resort Facilities"
                ],
                "summary": "Get semua fasilitas resort (Admin)",
                "operationId": "eb5780c66214992e2ae161ffe2ca1752",
                "responses": {
                    "200": {
                        "description": "Daftar fasilitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResortFacility"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Resort Facilities"
                ],
                "summary": "Membuat fasilitas resort baru",
                "operationId": "6033933c8167199efae5b4604d605334",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreResortFacilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fasilitas berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResortFacility"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/resort-facilities/{id}": {
            "get": {
                "tags": [
                    "Admin - Resort Facilities"
                ],
                "summary": "Get detail fasilitas (Admin)",
                "operationId": "9b3ef14c9ea4eacf026e023ab40fd607",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail fasilitas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResortFacility"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fasilitas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Resort Facilities"
                ],
                "summary": "Memperbarui fasilitas (Update)",
                "description": "Menggunakan POST untuk update karena PUT/PATCH bermasalah dengan form-data/file upload",
                "operationId": "2884c16282a2beb0525b738680ec6a22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResortFacilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fasilitas berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResortFacility"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fasilitas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Resort Facilities"
                ],
                "summary": "Menghapus fasilitas",
                "operationId": "ae5924c07d442e69e2ca634b69fe6fe2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fasilitas berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/superadmin/users": {
            "get": {
                "tags": [
                    "Super Admin - Users"
                ],
                "summary": "Get semua data pengguna (Super Admin)",
                "operationId": "1b9d16306b861109426982d43418d85b",
                "responses": {
                    "200": {
                        "description": "Daftar semua pengguna",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/superadmin/users/admin": {
            "post": {
                "tags": [
                    "Super Admin - Users"
                ],
                "summary": "Membuat akun Admin baru (Super Admin)",
                "operationId": "09b9b21c70a9bbf3ab0504139b80ea0d",
                "requestBody": {
                    "description": "Data akun admin baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Admin Baru"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "adminbaru@detuna.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "passwordadmin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Akun admin berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/superadmin/users/{id}": {
            "delete": {
                "tags": [
                    "Super Admin - Users"
                ],
                "summary": "Menghapus pengguna (Super Admin)",
                "operationId": "6c43f8ed96d7d1dfa2f2ddf0f292657d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengguna berhasil dihapus"
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreResortFacilityRequest": {
                "title": "Store Resort Facility Request",
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Nama Fasilitas Baru"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deskripsi fasilitas..."
                    },
                    "photo_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/image.jpg"
                    }
                },
                "type": "object"
            },
...
            "UpdateResortFacilityRequest": {
                "title": "Update Resort Facility Request",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Nama Fasilitas (Update)"
                    },
                    "description": {
                        "type": "string",
                        "example": "Deskripsi baru..."
                    }
                },
                "type": "object"
            },
...
            "ResortFacility": {
                "title": "Resort Facility",
                "description": "Model Fasilitas Umum Resort",
                "properties": {
                    "resort_facility_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Infinity Pool"
                    },
                    "description": {
                        "type": "string",
                        "example": "Kolam renang tanpa batas menghadap laut."
                    },
                    "photo_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/pool.jpg"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "Model Pengguna",
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "pelanggan",
                            "admin",
                            "super_admin"
                        ],
                        "example": "pelanggan"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "sanctum": {
            "type": "http",
            "description": "Masukkan Bearer Token (Sanctum)",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}